requirements.txt
streamlit
langchain
openai
chromadb
python-docx
pdfplumber
tiktoken